refactor: extract session notices from AgentWorkspace
